Move all EcalTrigPrimProducer references to cc file
